Almost works.

+++ source/demolib.h
@@ -24,7 +24,7 @@ void demolib_loadContext(lua_State* L, const char* s, int i);
 /// @param y Another arg.
 /// @param res The answer.
 /// @return Result.
-void demolib_luafunc_someCalc(lua_State* L, int x, int y, float* res);
+void demolib_luafunc_someCalc(lua_State* L, int x, int y, double* res);
 
 /// Call a lua function from C. Process a digital input change.
 /// @param L Lua environment.
